optimizations
Former-commit-id: 2fd719a121334f61b0985083e4885a5aa9b84210 [formerly dc89bd2d39860ea64cb297b1654539d5bed29bcf]
Former-commit-id: 9f86ba4ebe45ca236e9fa6b908aad772e87d2e7b

+++ Content/Texture.cpp
@@ -22,7 +22,7 @@ uint8_t* load(const string& filename, bool srgb, uint32_t& pixelSize, int32_t& x
 		srgb = false;
 	} else if (stbi_is_hdr(filename.c_str())) {
 		pixels = (uint8_t*)stbi_loadf(filename.c_str(), &x, &y, &channels, desiredChannels);
-		pixelSize = sizeof(uint32_t);
+		pixelSize = sizeof(float);
 		srgb = false;
 	} else {
 		pixels = (uint8_t*)stbi_load(filename.c_str(), &x, &y, &channels, desiredChannels);
